+++ src/6_Lights.h
@@ -1,9 +1,13 @@
 #include <Arduino.h>
 
-/* General LIGHTS SETTINGS ***********************************************************************************************************************
+/* LIGHTS SETTINGS ******************************************************************************************************************************************
  *
- * Most lights settings are done in the vehicle configuration files in the /vehicles/ directory.
- *
+ * All light settings are now done here and are not part of the vehicle file anymore
+ * All #define settings are only adjustable during compile time and are not changeable afterwards!
+ * 
+ * The other ones can be changed here or on 192.168.4.1, if connected to your vehicles WiFi access point
+ * Note, that the settings are only applied on a new or comletely erased ESP32. Afterwards they are stored in the EEPROM
+ * They can also be applied by changing the "eeprom_id" in "0_GeneralSettings"
  */
 
 // Neopixel settings ----------------------------------------------------------------------------------------------------------------------------------------
@@ -24,23 +28,22 @@
  * 4 = Union Jack United Kingdom animation for 8 LED (use it in combination with "BritishNationalAnthemSiren.h" sound in your vehicle file)
  * 5 = B33lz3bub Austria animation for 3 LED: https://www.rc-modellbau-portal.de/index.php?threads/baubericht-mercedes-actros-1851-gigaspace-tamiya.14349/page-3
  */
-uint8_t neopixelMode = 2;
+uint8_t neopixelMode = 2; // See above
 
-// These light options are adjustabale during compile time only ---------------------------------------------------------------------------------------------
-#define THIRD_BRAKELIGHT                 // if defined, pin 32 is used for a third brake light, otherwise for a trailer presence switch (pulled to GND = no trailer attached)
-#define HAZARDS_WHILE_5TH_WHEEL_UNLOCKED // Hazards on, if 5th wheel is unlocked
+// These light settings are adjustabale during compile time only ---------------------------------------------------------------------------------------------
+#define THIRD_BRAKELIGHT // if defined, pin 32 is used for a third brake light, otherwise for a trailer presence switch (pulled to GND = no trailer attached)
 
 // These light options are adjustable on the configuration website and stored in the EEPROM -----------------------------------------------------------------
-bool noCabLights = false;             // The cablights step in the lights sequence is skipped, if true
-bool noFogLights = false;             // The foglights step in the lights sequence is skipped, if true
-bool xenonLights = false;             // Headlights will show a xenon bulb ignition flash, if true
-bool flickeringWileCranking = false;  // if true: The lights will flicker a bit during engine cranking, otherwise they are just a bit dimmed
-bool ledIndicators = false;           // indicators are switching "hard" (without fading like an incandescent bulb), if true (was LED_INDICATORS)
-bool swap_L_R_indicators = false;     // If you want to swap L & R indicators
-bool indicatorsAsSidemarkers = false; // The indicators are used as side markers as well. This is commonly used in US trucks.
-bool separateFullBeam = false;        // The full beam is a separate bulb, connected to roof lights pin. Connect roof lights to side lights in this case
-
-bool flashingBlueLight = true; // double flash blue lights if "true", "rotating" beacons if "false"
+bool noCabLights = false;                 // The cablights step in the lights sequence is skipped, if true
+bool noFogLights = false;                 // The foglights step in the lights sequence is skipped, if true
+bool xenonLights = false;                 // Headlights will show a xenon bulb ignition flash, if true
+bool flickeringWileCranking = false;      // if true: The lights will flicker a bit during engine cranking, otherwise they are just a bit dimmed
+bool ledIndicators = false;               // indicators are switching "hard" (without fading like an incandescent bulb), if true (was LED_INDICATORS)
+bool swap_L_R_indicators = false;         // If you want to swap L & R indicators
+bool indicatorsAsSidemarkers = false;     // "US Mode" The indicators are used as side markers as well. This is commonly used in US trucks.
+bool separateFullBeam = false;            // The full beam is a separate bulb, connected to roof lights pin. Connect roof lights to side lights in this case
+bool flashingBlueLight = true;            // double flash blue lights if "true", "rotating" beacons if "false"
+bool hazardsWhile5thWheelUnlocked = true; // Hazards on, if 5th wheel is unlocked
 
 // Light brightness values 0 - 255
 uint8_t cabLightsBrightness = 100;      // Usually 255, 100 for Actros & Ural